requirements.txt
aiohttp==3.7.4.post0
amqp==2.6.1
appdirs==1.4.4
asgiref==3.2.7
astroid==2.4.2
async-timeout==3.0.1
attrs==21.2.0
auditlog3==1.0.1
autoflake==1.4
autopep8==1.5.3
Babel==2.8.0
beautifulsoup4==4.9.1
billiard==3.6.4.0
black==20.8b1
boto3==1.17.58
botocore==1.20.69
cachetools==4.2.1
celery==4.4.7
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
defusedxml==0.6.0
diff-match-patch==20181111
Django==3.0.7
django-admin-interface==0.14.2
django-appconf==1.0.4
django-bootstrap-form==3.4
django-ckeditor==5.9.0
django-colorfield==0.4.0
django-compressor==2.4.1
django-crispy-forms==1.11.1
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-gsheets==0.0.10
django-import-export==2.2.0
django-jazzmin==2.4.4
django-js-asset==1.2.2
django-jsonfield==1.4.1
django-letsencrypt==4.0.0
django-mail-templated==2.6.5
django-multiselectfield==0.1.12
django-phone-field==1.8.1
django-phonenumber-field==4.0.0
django-slick-reporting==0.5.7
django-storages==1.11.1
django-utils-six==2.0
docutils==0.15.2
et-xmlfile==1.0.1
flake8==3.8.4
flake8-black==0.2.1
google-api-core==1.26.0
google-api-python-client==1.12.8
google-auth==1.27.0
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.2
googleapis-common-protos==1.52.0
gunicorn==20.0.4
httplib2==0.19.0
idna==2.10
importlib-metadata==3.4.0
ipinfo==4.2.0
isort==4.3.21
jdcal==1.4.1
jmespath==0.10.0
kombu==4.6.11
lazy-object-proxy==1.4.3
MarkupPy==1.14
mccabe==0.6.1
multidict==5.1.0
mypy-extensions==0.4.3
oauthlib==3.1.0
odfpy==1.4.1
openpyxl==3.0.3
orjson==3.4.8
packaging==20.9
pathspec==0.8.1
phonenumbers==8.12.5
Pillow==8.2.0
protobuf==3.15.1
psycopg2-binary==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pyflakes==2.2.0
pylint==2.5.3
pylint-django==2.0.15
pylint-plugin-utils==0.6
pyparsing==2.4.7
python-dateutil==2.8.1
python-decouple==3.3
pytz==2020.1
PyYAML==5.3.1
rcssmin==1.0.6
regex==2020.11.13
requests==2.24.0
requests-oauthlib==1.3.0
rjsmin==1.1.0
rsa==4.7.1
s3transfer==0.4.2
simplejson==3.17.2
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
structlog==21.1.0
tablib==2.0.0
toml==0.10.1
typed-ast==1.4.1
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.25.9
vine==1.3.0
whitenoise==5.1.0
wrapt==1.12.1
xlrd==1.2.0
xlwt==1.3.0
yarl==1.6.3
zipp==3.4.0